Initial workings for a preferences file
The preferences file will be used to hold the enabled and disabled state
of repos. Eventually, these states will be able to be toggled through
the mirror.sh program.

Currently, `mirror.sh -l` is needed to generate the initial list of
preferences. Only the generation of the list is implemented.